Поправил характеристики, начал работу над UI, исправил некоторые баги, ввел систему брони

+++ Assets/Resources/StatusData.cs
@@ -23,24 +23,24 @@ public class StatusData : ScriptableObject
 
     [Header("Основнык характеристики")]
     //Сила
-    [SerializeField, Range(-100f, 100f)] public int strength = 0;
-    public float Strength { get => strength; }
+    [SerializeField, Range(-100, 100)] public int strength = 0;
+    public int Strength { get => strength; }
 
     //Ловкость
-    [SerializeField, Range(-100f, 100f)] public int agility = 0;
-    public float Agility { get => agility; }
+    [SerializeField, Range(-100, 100)] public int agility = 0;
+    public int Agility { get => agility; }
 
     //Интеллект
-    [SerializeField, Range(-100f, 100f)] public int intel = 0;
-    public float Intel { get => intel; }
+    [SerializeField, Range(-100, 100)] public int intel = 0;
+    public int Intel { get => intel; }
 
     //Телосложение
-    [SerializeField, Range(-100f, 100f)] public int constitution = 0;
-    public float Constitution { get => constitution; }
+    [SerializeField, Range(-100, 100)] public int constitution = 0;
+    public int Constitution { get => constitution; }
 
     //Мудрость
-    [SerializeField, Range(-100f, 100f)] public int wisdom = 0;
-    public float Wisdom { get => wisdom; }
+    [SerializeField, Range(-100, 100)] public int wisdom = 0;
+    public int Wisdom { get => wisdom; }
 
 
 
@@ -48,36 +48,36 @@ public class StatusData : ScriptableObject
 
     [Header("Вторичные характеристики")]
     //Уклонение
-    [SerializeField, Range(-100f, 100f)] public int dodge = 0;
-    public float Dodge { get => dodge; }
+    [SerializeField, Range(-100, 100)] public int dodge = 0;
+    public int Dodge { get => dodge; }
 
     //Переносимый вес
-    [SerializeField, Range(-100f, 100f)] public int carryingCapacity = 0;
-    public float CarryingCapacity { get => carryingCapacity; }
+    [SerializeField, Range(-100, 100)] public int carryingCapacity = 0;
+    public int CarryingCapacity { get => carryingCapacity; }
 
     //Скорость
-    [SerializeField, Range(-100f, 100f)] public float speed = 0;
-    public float Speed { get => speed; }
+    [SerializeField, Range(-100, 100)] public int speed = 0;
+    public int Speed { get => speed; }
 
     //Скорость атаки
-    [SerializeField, Range(-100f, 100f)] public float attackSpeed = 0;
-    public float AttackSpeed { get => attackSpeed; }
+    [SerializeField, Range(-100, 100)] public int attackSpeed = 0;
+    public int AttackSpeed { get => attackSpeed; }
 
     //Критический урон
-    [SerializeField, Range(-100f, 100f)] public float criticalDamage = 0;
-    public float CriticalDamage { get => criticalDamage; }
+    [SerializeField, Range(-100, 100)] public int criticalDamage = 0;
+    public int CriticalDamage { get => criticalDamage; }
 
     //Шанс критануть
-    [SerializeField, Range(-100f, 100f)] public float criticalDamageChance = 0;
-    public float CriticalDamageChance { get => criticalDamageChance; }
+    [SerializeField, Range(-100, 100)] public int criticalDamageChance = 0;
+    public int CriticalDamageChance { get => criticalDamageChance; }
 
     //Точность
-    [SerializeField, Range(-100f, 100f)] public float precision = 0;
-    public float Precision { get => precision; }
+    [SerializeField, Range(-100, 100)] public int precision = 0;
+    public int Precision { get => precision; }
 
     //Опьянение
-    [SerializeField, Range(-100f, 100f)] public int drunkenness = 0;
-    public float Drunkenness { get => drunkenness; }
+    [SerializeField, Range(-100, 100)] public int drunkenness = 0;
+    public int Drunkenness { get => drunkenness; }
 
 
 
@@ -85,61 +85,61 @@ public class StatusData : ScriptableObject
 
     [Header("Сопротивления к урону")]
     //Сопротивление колющему📌
-    [SerializeField, Range(-100f, 100f)] public float prickResist = 0;
-    public float PrickResist { get => prickResist; }
+    [SerializeField, Range(-100, 100)] public int prickResist = 0;
+    public int PrickResist { get => prickResist; }
 
     //Сопротивление режущему🔪
-    [SerializeField, Range(-100f, 100f)] public float slashResist = 0;
-    public float SlashResist { get => slashResist; }
+    [SerializeField, Range(-100, 100)] public int slashResist = 0;
+    public int SlashResist { get => slashResist; }
 
     //Сопротивление дробящему🔨
-    [SerializeField, Range(-100f, 100f)] public float crushResist = 0;
-    public float CrushResist { get => crushResist; }
+    [SerializeField, Range(-100, 100)] public int crushResist = 0;
+    public int CrushResist { get => crushResist; }
 
     //Сопротивление ядам🍄
-    [SerializeField, Range(-100f, 100f)] public float poisonResist = 0;
-    public float PoisonResist { get => poisonResist; }
+    [SerializeField, Range(-100, 100)] public int poisonResist = 0;
+    public int PoisonResist { get => poisonResist; }
 
     //Сопротивление огню🔥
-    [SerializeField, Range(-100f, 100f)] public float fireResist = 0;
-    public float FireResist { get => fireResist; }
+    [SerializeField, Range(-100, 100)] public int fireResist = 0;
+    public int FireResist { get => fireResist; }
 
     //Сопростивление морозу❄ 
-    [SerializeField, Range(-100f, 100f)] public float frostResist = 0;
-    public float FrostResist { get => frostResist; }
+    [SerializeField, Range(-100, 100)] public int frostResist = 0;
+    public int FrostResist { get => frostResist; }
 
     //Сопротивление проклятию☠
-    [SerializeField, Range(-100f, 100f)] public float curseResist = 0;
-    public float CurseResist { get => curseResist; }
+    [SerializeField, Range(-100, 100)] public int curseResist = 0;
+    public int CurseResist { get => curseResist; }
 
     //Сопротивление электричеству⛈
-    [SerializeField, Range(-100f, 100f)] public float electricalResist = 0;
-    public float ElectricalResist { get => electricalResist; }
+    [SerializeField, Range(-100, 100)] public int electricalResist = 0;
+    public int ElectricalResist { get => electricalResist; }
 
     //Сопротивление АлКоГоЛю🍺
-    [SerializeField, Range(-100f, 100f)] public float drunkennessResist = 0;
-    public float DrunkennessResist { get => drunkennessResist; }
+    [SerializeField, Range(-100, 100)] public int drunkennessResist = 0;
+    public int DrunkennessResist { get => drunkennessResist; }
 
 
 
     [Header("Сопротивления к урону")]
     //Сопротивление АлКоГоЛю🍺
-    [SerializeField, Range(-100f, 100f)] public int hammerWay = 0;
+    [SerializeField, Range(-100, 100)] public int hammerWay = 0;
     public int HammerWay { get => hammerWay; }
 
     //Сопротивление АлКоГоЛю🍺
-    [SerializeField, Range(-100f, 100f)] public int gearWay = 0;
+    [SerializeField, Range(-100, 100)] public int gearWay = 0;
     public int GearWay { get => gearWay; }
 
     //Сопротивление АлКоГоЛю🍺
-    [SerializeField, Range(-100f, 100f)] public int anvilWay = 0;
+    [SerializeField, Range(-100, 100)] public int anvilWay = 0;
     public int AnvilWay { get => anvilWay; }
 
     //Сопротивление АлКоГоЛю🍺
-    [SerializeField, Range(-100f, 100f)] public int beerWay = 0;
+    [SerializeField, Range(-100, 100)] public int beerWay = 0;
     public int BeerWay { get => beerWay; }
 
     //Сопротивление АлКоГоЛю🍺
-    [SerializeField, Range(-100f, 100f)] public int runeWay = 0;
+    [SerializeField, Range(-100, 100)] public int runeWay = 0;
     public int RuneWay { get => runeWay; }
 }
